+++ docs/changelog.rst
@@ -88,7 +88,7 @@ statistics (:py:class:`yasa.SleepStatsAgreement`):
 **New functions**
 
 * New :py:func:`yasa.fetch_sample` function to download and cache sample YASA data files. (`PR 192 <https://github.com/raphaelvallat/yasa/pull/192>`_)
-* New :py:meth:`yasa.Hypnogram.from_integers` classmethod to construct a :py:class:`yasa.Hypnogram` directly from a legacy integer-encoded array (e.g. loaded from a plain-text file with :py:func:`numpy.loadtxt`).
+* New :py:meth:`yasa.Hypnogram.from_integers` classmethod to construct a :py:class:`yasa.Hypnogram` directly from a legacy integer-encoded array.
 
 **API changes**